test(packages/lint-repository-sui): move tests to server
