+++ packages/lint-repository-sui/test/server/rules/node-versionSpec.js
@@ -1,7 +1,7 @@
 import dedent from 'dedent'
 
 import handler from '../../src/rules/node-version.js'
-import {MatchStub, RuleTester} from '../TestHelpers'
+import {MatchStub, RuleTester} from '../../TestHelpers.js'
 
 RuleTester.create('node-version', handler).run({
   valid: [
