+++ packages/lint-repository-sui/test/server/rules/package-lockSpec.js
@@ -1,7 +1,7 @@
 import dedent from 'dedent'
 
 import handler from '../../src/rules/package-lock.js'
-import {RuleTester} from '../TestHelpers'
+import {RuleTester} from '../../TestHelpers.js'
 
 RuleTester.create('package-lock', handler).run({
   valid: [
