+++ packages/lint-repository-sui/test/server/rules/react-versionSpec.js
@@ -1,7 +1,7 @@
 import dedent from 'dedent'
 
 import handler from '../../src/rules/react-version.js'
-import {MatchStub, RuleTester} from '../TestHelpers'
+import {MatchStub, RuleTester} from '../../TestHelpers.js'
 
 RuleTester.create('react-version', handler).run({
   valid: [
